test Seeding

+++ src/KidsToyHive.Domain/Features/Products/ProductDto.cs
@@ -18,7 +18,7 @@ namespace KidsToyHive.Domain.Features.Products
     {        
         public Guid ProductId { get; set; }
         public string Name { get; set; }
-        public float HourlyRate { get; set; }
+        public int HourlyRate { get; set; }
         public int Version { get; set; }
     }
